Extract `getUrl` to utils
